Fix query handling in epo id lookup
In a previous commit, the response handling for the epo id service
registry query was changed to use MessageUtils.json_payload_to_dict
to decode the response. This is incorrect, however, in that the response
is actually a string and not a DXL message, causing an exception to be
thrown when the id lookup is complete. In this commit, the
json_to_dict() method is used instead, which handles the response
properly and avoids the exception.