Added ability to use DXL 5 ePO commands service
This commit adds the ability for the ePO Python client to be able to
make remote commands to the ePO 'commands' service, which is being added
to the DXL ePO extensions for DXL 5. This commit retains the ability to
use the ePO 'remote' service, which is hosted by the standalone DXL ePO
Python service, for making remote commands. When an ePO unique id is not
provided to the client, the client will search for an ePO 'commands'
service, falling back to looking for a 'remote' service if no 'commands'
service can be found.

An Exception is raised if an attempt is made to perform a remote command
with a non-JSON output format when the target service is a DXL
'commands' service, since this service only supports the JSON output
format. The help() method, however, can still be called with an output
format of 'verbose'. In this case, the output format requested from the
server is JSON. The method flattens the return array into plain text,
delimited by line separators, for backward compatibility.