Combine commands and remote service ids together during lookup
Previously, a `lookup_epo_unique_identifiers` call would return
either the list of `remote` service ids (if at least one `remote`
service was available) or the list of `commands` service ids (if no
`remote` services were available). The returned list would not
include the superset of available `remote` and `commands` service
ids.

In this commit, the combined list of `remote` and `commands` service
ids is returned for a `lookup_epo_unique_identifiers` call.